[support/network] Add Windows support
